Fix lint issues take 2
* errcheck: 50
* goconst: 1
* goimports: 3
* gosec: 50
* ineffassign: 2
* misspell: 2
* prealloc: 17
* staticcheck: 45
* unconvert: 1

+++ sei-tendermint/internal/p2p/address.go
@@ -77,7 +77,7 @@ func ParseNodeAddress(urlString string) (NodeAddress, error) {
 	}
 	// For some reasons, missing or 0 port on parsing is interpretented as the default port.
 	if address.Port == 0 {
-		address.Port = uint16(defaultPort)
+		address.Port = defaultPort
 	}
 	return address, address.Validate()
 }
